Maryia/isolated PlaygroundComponent from Playground page, used sessionStorage instead of localStorage
