+++ src/components/ConsoleMessage/ConsoleMessage.tsx
@@ -1,14 +1,14 @@
-import React from "react";
 import CodeContent from "components/common/CodeContent/CodeContent";
-import style from "../../pages/playground/Playground.module.scss";
+import React from "react";
+import style from "components/PlaygroundComponent/PlaygroundComponent.module.scss";
 
 const ConsoleMessage: React.FC<{ message: any }> = ({ message }) => {
-    const payload = JSON.stringify(message.body, null, 4)
+    const payload = JSON.stringify(message.body, null, 4);
     return (
         <div className={style[message.type]}>
             <CodeContent lang="json" data={payload}></CodeContent>
         </div>
-    )
-}
+    );
+};
 
 export default ConsoleMessage;
