+++ src/components/PlaygroundComponent/PlaygroundComponent.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors.module";
+@import "styles/colors.module";
 
 .playground-content {
     padding: 180px 5px;
